+++ chicago-web/src/views/Dashboard/Dashboard.js
@@ -20,11 +20,10 @@ import {
   Row,
   Table,
 } from 'reactstrap';
-import Widget03 from '../../views/Widgets/Widget03';
+import Widget03 from '../../views/Widgets/Widget03'
 import { CustomTooltips } from '@coreui/coreui-plugin-chartjs-custom-tooltips';
-import { getStyle, hexToRgba } from '@coreui/coreui/dist/js/coreui-utilities';
+import { getStyle, hexToRgba } from '@coreui/coreui-pro/dist/js/coreui-utilities'
 import config from 'react-global-configuration';
-
 
 const brandPrimary = getStyle('--primary')
 const brandSuccess = getStyle('--success')
@@ -194,7 +193,7 @@ const cardChartOpts3 = {
 
 // Card Chart 4
 const cardChartData4 = {
-  labels: ['', '', '', '', '', '', '', '', '', '', '', '', '', '', '', ''],
+  labels: ['January', 'February', 'March', 'April', 'May', 'June', 'July', 'August', 'September', 'October', 'November', 'December', 'January', 'February', 'March', 'April'],
   datasets: [
     {
       label: 'My First dataset',
@@ -516,10 +515,10 @@ class Dashboard extends Component {
                 </ButtonGroup>
                 <div className="text-value">9.823</div>
                 <div>Members online</div>
+                <div className="chart-wrapper mt-3" style={{ height: '70px' }}>
+                  <Line data={cardChartData2} options={cardChartOpts2} height={70} />
+                </div>
               </CardBody>
-              <div className="chart-wrapper mx-3" style={{ height: '70px' }}>
-                <Line data={cardChartData2} options={cardChartOpts2} height={70} />
-              </div>
             </Card>
           </Col>
 
@@ -540,10 +539,10 @@ class Dashboard extends Component {
                 </ButtonGroup>
                 <div className="text-value">9.823</div>
                 <div>Members online</div>
+                <div className="chart-wrapper mt-3" style={{ height: '70px' }}>
+                  <Line data={cardChartData1} options={cardChartOpts1} height={70} />
+                </div>
               </CardBody>
-              <div className="chart-wrapper mx-3" style={{ height: '70px' }}>
-                <Line data={cardChartData1} options={cardChartOpts1} height={70} />
-              </div>
             </Card>
           </Col>
 
@@ -565,7 +564,7 @@ class Dashboard extends Component {
                 <div className="text-value">9.823</div>
                 <div>Members online</div>
               </CardBody>
-              <div className="chart-wrapper" style={{ height: '70px' }}>
+              <div className="chart-wrapper mt-3" style={{ height: '70px' }}>
                 <Line data={cardChartData3} options={cardChartOpts3} height={70} />
               </div>
             </Card>
@@ -589,7 +588,7 @@ class Dashboard extends Component {
                 <div className="text-value">9.823</div>
                 <div>Members online</div>
               </CardBody>
-              <div className="chart-wrapper mx-3" style={{ height: '70px' }}>
+              <div className="chart-wrapper mt-3 mx-3" style={{ height: '70px' }}>
                 <Bar data={cardChartData4} options={cardChartOpts4} height={70} />
               </div>
             </Card>